requirements.txt
flask==2.3.2
RPi.GPIO==0.7.1
gpiozero==1.6.2